Move `company/accounts.html`
Move template to `user/company_accounts.html`. Rename some modules and classes.
